+++ components/shared/AppSidebar/InsightsPanel.tsx
@@ -63,7 +63,7 @@ export const InsightsPanel = ({ title, username, insights, type, isLoading, work
                     ? `/workspaces/${workspaceId}/contributor-insights/${insight.id}/overview`
                     : `/lists/${insight.id}`
                   : workspaceId
-                  ? `/workspaces/${workspaceId}/repository-insights/${insight.id}/overview`
+                  ? `/workspaces/${workspaceId}/repository-insights/${insight.id}/dashboard`
                   : `/pages/${username}/${insight.id}/dashboard`;
               return (
                 <li