+++ middleware.ts
@@ -2,7 +2,6 @@ import { createMiddlewareClient } from "@supabase/auth-helpers-nextjs";
 import { NextResponse } from "next/server";
 import { NextRequest } from "next/server";
 import { pathToRegexp } from "path-to-regexp";
-import { getAllFeatureFlags } from "lib/utils/server/feature-flags";
 import { WORKSPACE_ID_COOKIE_NAME, getWorkspaceUrl } from "lib/utils/workspace-utils";
 
 // HACK: this is to get around the fact that the normal next.js middleware is not always functioning
@@ -72,22 +71,14 @@ export async function middleware(req: NextRequest) {
     return res;
   }
 
-  // TODO: remove this once we've rolled this out to everyone.
-  // For now, only allowed users with the workspaces feature flag can access the workspaces pages.
   if (session?.user && req.nextUrl.pathname.startsWith("/workspaces")) {
-    const featureFlags = await getAllFeatureFlags(Number(session?.user.user_metadata.sub));
+    const [, , workspaceId] = req.nextUrl.pathname.split("/");
 
-    if (featureFlags.workspaces) {
-      const [, , workspaceId] = req.nextUrl.pathname.split("/");
-
-      if (workspaceId !== "new") {
-        res.cookies.set(WORKSPACE_ID_COOKIE_NAME, workspaceId);
-      }
-
-      return res;
-    } else {
-      return NextResponse.rewrite(new URL("/404", req.url));
+    if (workspaceId !== "new") {
+      res.cookies.set(WORKSPACE_ID_COOKIE_NAME, workspaceId);
     }
+
+    return res;
   }
 
   // Check auth condition
@@ -97,15 +88,9 @@ export async function middleware(req: NextRequest) {
       const data = await loadSession(req, session?.access_token);
 
       if (data.is_onboarded) {
-        const featureFlags = await getAllFeatureFlags(Number(session?.user.user_metadata.sub));
+        const workspaceUrl = getWorkspaceUrl(req.cookies, req.url, data.personal_workspace_id);
 
-        if (featureFlags.workspaces) {
-          const workspaceUrl = getWorkspaceUrl(req.cookies, req.url, data.personal_workspace_id);
-
-          return NextResponse.redirect(`${workspaceUrl}`);
-        } else {
-          return NextResponse.redirect(new URL("/hub/insights", req.url));
-        }
+        return NextResponse.redirect(`${workspaceUrl}`);
       } else {
         return NextResponse.redirect(new URL("/feed", req.url));
       }